+++ includes/build.php
@@ -1,5 +1,5 @@
 <?php
 if (!defined('ABSPATH')) { exit; }
-define('SLTR_BUILD_VERSION', '1.0.1046');
-define('SLTR_BUILD_ID', '1.0.1046');
+define('SLTR_BUILD_VERSION', '1.0.1047');
+define('SLTR_BUILD_ID', '1.0.1047');